fix test in test file
